pyproject.toml
[project]
name = "finviz-screener"
version = "0.1.0"
description = "Scheduled momentum stock screener powered by Finviz + Claude"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.49",
    "httpx>=0.28",
    "anthropic>=0.40",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "typer>=0.15",
    "fastapi>=0.115",
    "uvicorn>=0.34",
]

[project.scripts]
finviz-screener = "finviz_screener.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/finviz_screener"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
# Long lines are inside the LLM system-prompt string — can't wrap prose.
"src/finviz_screener/analyzer.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
